refactor: split storage/memory.rs into focused sub-modules (#144)
